+++ eurocropsml/acquisition/clipping/clipper.py
@@ -163,10 +163,8 @@ def _filter_args(
 
 def _process_raster_parallel(
     satellite: Literal["S1", "S2"],
-    denoise: bool,
     polygon_df: pd.DataFrame,
     parcel_id_name: str,
-    bands: list[str],
     filtered_images: gpd.GeoDataFrame,
     band_tiles: list[Path],
 ) -> pd.DataFrame:
@@ -174,12 +172,9 @@ def _process_raster_parallel(
 
     Args:
         satellite: S1 for Sentinel-1 and S2 for Sentinel-2.
-        denoise: Whether to perform thermal noise removal for Sentinel-1.
-            For Sentinel-2 this argument has no effect.
         polygon_df: Dataframe containing all parcel ids. Will be merged with the clipped values.
         parcel_id_name: The country's parcel ID name (varies from country to country).
         filtered_images: Dataframe containing all parcel ids that lie in this raster tile.
-        bands: (Sub-)set of Sentinel-1 (radar) or Sentinel-2 (spectral) bands.
         band_tiles: Paths to the raster's band tiles.
 
     Returns:
@@ -197,7 +192,7 @@ def _process_raster_parallel(
         filtered_geom = polygon_df[polygon_df[parcel_id_name].isin(parcel_ids)]
 
         result = mask_polygon_raster(
-            satellite, band_tiles, bands, filtered_geom, parcel_id_name, product_date, denoise
+            satellite, band_tiles, filtered_geom, parcel_id_name, product_date
         )
 
         if result is not None:
@@ -255,10 +250,8 @@ def clipping(
     func = partial(
         _process_raster_parallel,
         config.satellite,
-        config.denoise,
         polygon_df,
         cast(str, config.parcel_id_name),
-        config.bands,
     )
 
     polygon_df = polygon_df.drop(["geometry"], axis=1)
@@ -267,6 +260,9 @@ def clipping(
 
     new_data: bool = False
     if processed < len(args):
+        if config.satellite == "S1":
+            # needed for esa_snappy
+            mp_orig.set_start_method("spawn", force=True)
         new_data = True
         logger.info("Starting parallel raster clipping...")
         te = tqdm(total=len(args) - processed, desc="Clipping raster tiles.")